[libs] E: Move Binding for send()
This commits moves the binding for the send() system calls to the
syscall crate.
